requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.5=py310hd125d64_0
aiosignal=1.3.1=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
aws-c-auth=0.7.22=haa5a189_7
aws-c-cal=0.7.0=h94d0942_0
aws-c-common=0.9.23=h99b78c6_0
aws-c-compression=0.2.18=h94d0942_7
aws-c-event-stream=0.4.2=he43e89f_14
aws-c-http=0.8.2=h638bd1a_4
aws-c-io=0.14.9=hf0e3e08_5
aws-c-mqtt=0.10.4=h80c1ce3_7
aws-c-s3=0.6.0=h11f64d3_0
aws-c-sdkutils=0.1.16=h94d0942_3
aws-checksums=0.1.18=h94d0942_7
aws-crt-cpp=0.27.2=hd9c8ee4_0
aws-sdk-cpp=1.11.329=haf867cf_8
azure-core-cpp=1.12.0=hd01fc5c_0
azure-identity-cpp=1.8.0=h0a11218_1
azure-storage-blobs-cpp=12.11.0=h77cc766_1
azure-storage-common-cpp=12.6.0=h7024f69_1
azure-storage-files-datalake-cpp=12.10.0=h64d02d0_1
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py310h1253130_1
bzip2=1.0.8=h93a5062_5
c-ares=1.32.0=h99b78c6_0
ca-certificates=2024.7.4=hf0a4a13_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.16.0=py310hdcd7c05_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=py310h21239e6_0
cycler=0.12.1=pyhd8ed1ab_0
datasets=2.20.0=pyhd8ed1ab_0
debugpy=1.8.2=py310hcf9f62a_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.15.4=pyhd8ed1ab_0
fonttools=4.53.1=py310ha6dd24b_0
freetype=2.12.1=hadb7bae_2
frozenlist=1.4.1=py310hd125d64_0
fsspec=2024.5.0=pyhff2d567_0
gflags=2.2.2=hc88da5d_1004
glog=0.7.1=heb240a5_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py310h3bc658a_1
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
huggingface_hub=0.23.4=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=hc8870d7_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hbe9552e_0
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.2=cxx17_hebf3989_0
libarrow=16.1.0=h71e69af_13_cpu
libarrow-acero=16.1.0=h00cdb27_13_cpu
libarrow-dataset=16.1.0=h00cdb27_13_cpu
libarrow-substrait=16.1.0=hc68f6b8_13_cpu
libblas=3.9.0=22_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=22_osxarm64_openblas
libcrc32c=1.1.2=hbdafb3b_0
libcurl=8.8.0=h7b6f9a7_1
libcxx=18.1.8=h167917d_0
libdeflate=1.20=h93a5062_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h2757513_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libgoogle-cloud=2.26.0=hfe08963_0
libgoogle-cloud-storage=2.26.0=h1466eeb_0
libgrpc=1.62.2=h9c18a4f_0
libiconv=1.17=h0d3ecfb_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.27=openmp_h517c56d_1
libparquet=16.1.0=hcf52c46_13_cpu
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libre2-11=2023.09.01=h7b2c953_2
libsodium=1.0.18=h27ca646_1
libsqlite=3.46.0=hfb93653_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.19.0=h026a170_1
libtiff=4.6.0=h07db509_3
libtorch=2.3.1=cpu_generic_hf1facdc_1
libutf8proc=2.8.0=h1a8c8d9_0
libuv=1.48.0=h93a5062_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.16=hf2054a2_0
libxml2=2.12.7=ha661575_1
libzlib=1.3.1=hfb2fe0b_1
llvm-openmp=18.1.8=hde57baf_0
lz4-c=1.9.4=hb7217d7_0
markupsafe=2.1.5=py310hd125d64_0
matplotlib=3.9.1=py310hb6292c7_0
matplotlib-base=3.9.1=py310heb73f16_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h41d338b_1
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.5=py310h8e9501a_0
multiprocess=0.70.16=py310hd125d64_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=hb89a1cb_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pyhd8ed1ab_1
nomkl=1.0=h5ca1d4c_0
numpy=1.26.4=py310hd45542a_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.1=hfb2fe0b_1
orc=2.0.1=h47ade37_1
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py310h2216879_1
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py310h01af8b1_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py310ha6dd24b_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=16.1.0=py310h24597f5_4
pyarrow-core=16.1.0=py310h2e300fa_4_cpu
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.14=h2469fbe_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python-xxhash=3.4.1=py310h2aa6e3c_0
python_abi=3.10=4_cp310
pytorch=2.3.1=cpu_generic_py310hb190f2a_1
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py310h2aa6e3c_1
pyzmq=26.0.3=py310h16e08c9_0
qhull=2020.2=h420ef59_3
re2=2023.09.01=h4cba328_2
readline=8.2=h92ec313_1
regex=2024.5.15=py310ha6dd24b_0
requests=2.32.3=pyhd8ed1ab_0
safetensors=0.4.3=py310hd442715_0
setuptools=70.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
snappy=1.2.1=hd02b534_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12.1=pypyh2585a3b_103
tiktoken=0.7.0=py310hdeb0604_0
tk=8.6.13=h5083fa2_1
tokenizers=0.19.1=py310hb145503_0
torchvision=0.18.1=py310_cpu
tornado=6.4.1=py310ha6dd24b_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.42.3=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py310h2aa6e3c_0
urllib3=2.2.2=pyhd8ed1ab_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xxhash=0.8.2=hb547adb_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.4=py310hd125d64_0
zeromq=4.3.5=hcc0f68c_4
zipp=3.19.2=pyhd8ed1ab_0
zstandard=0.22.0=py310hd9c37c8_1
zstd=1.5.6=hb46c0d2_0
